Update Poly_thesis.ipynb
TestOdClustering function is unified with OdClustering function .The shuffling functionality moved to the latter and kicks in only when needed (shuffle=True or either set of k values for start or end points has more than 1 element).
Also start and end points are now plotted independently to avoid rainsing an error if the set of "k" values for start and end points ("nClusOriginSet" & "nClusDestSet") are not of the same size.